examples/blob-and-file-triggers/blob_thumbnail_generator/requirements.txt
azure-functions==1.21.3
azure-functions-logging==0.5.0
azure-storage-blob==12.25.1
Pillow==11.2.1
